Make setOkHttpClientDefaults public
With the change to Retrofit 2 tmdb-java's Interceptors must
now be added to the OkHttpClient prior to it being built. This
makes it impossible for a user of tmdb-java to provide their own
custom OkHttpClient because it won't have the Interceptor required
to add the API key to the requests (causing all requests to be refused
by TMDB)

Make setOkHttpClientDefaults public instead of protected so that users
of tmdb-java can add the interceptors to their custom OkHttp client
builder prior to building it.